installer/linux: run service under non-root service account
Update the linux installer scripts to create a low privileged,
non-interactive, service account on the system.

Update the `sample-trace2-otel-collector.service` to tell
`systemctl` to run the collector service under our new service
account.

Chown a few of the directories and files in the installation
directory so that the service can function with these reduced
privileges. This is just enough to let the service create the
Unix domain socket and write the various install/stdout/stderr
logs.  The remainder of the installation is still owned by the
installing user or root.

The user name "trace2" is used by default.

Additionally, updated the various scripts to remove the `set -e`
option. This was causing difficulties for for `dpkg` when we
actually wanted to use the exit code for branching within the
scripts.

Signed-off-by: Jeff Hostetler <jeffhostetler@github.com>